settings and arabic strings

+++ Not_Today_Sun/app/src/main/java/com/example/not_today_sun/notification/view/AlarmAdapter.kt
@@ -2,6 +2,7 @@ package com.example.not_today_sun.notification.view
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
@@ -49,9 +50,21 @@ class AlarmAdapter(
                     false -> "Notifications Enabled"
                 }
                 buttonDelete.setOnClickListener {
-                    onDeleteClick(alarm)
-                }
+                    AlertDialog.Builder(binding.root.context)
+                        .setTitle("Delete Location")
+                        .setMessage("Are you sure you want to delete item ?")
+                        .setPositiveButton("Yes") { _, _ ->
+                            onDeleteClick(alarm)
+                        }
+                        .setNegativeButton("No") { dialog, _ ->
+                            dialog.dismiss()
+                        }
+                        .setCancelable(true)
+                        .show()                }
+
+
+
+            }
             }
         }
     }
-}
